refactor(api): all projects and search
remove redundant search code, replace with all projects and starting to conform to REST principles
