+++ src/lib/server/utils/projects.js
@@ -1,10 +1,10 @@
-import { supabase } from "$lib/server/supabase.js";
+import { supabase } from '$lib/server/supabase.js';
 
 export async function fetchProjectCategories(projectIds) {
   const { data: projectCategories, error } = await supabase
-    .from("category_project")
-    .select("project_id, category_id")
-    .in("project_id", projectIds);
+    .from('category_project')
+    .select('project_id, category_id')
+    .in('project_id', projectIds);
 
   if (error) throw new Error(error.message);
   return projectCategories;
@@ -12,9 +12,9 @@ export async function fetchProjectCategories(projectIds) {
 
 export async function fetchCategories(categoryIds) {
   const { data: categories, error } = await supabase
-    .from("categories")
-    .select("*")
-    .in("id", categoryIds);
+    .from('categories')
+    .select('*')
+    .in('id', categoryIds);
 
   if (error) throw new Error(error.message);
   return categories;
@@ -22,20 +22,15 @@ export async function fetchCategories(categoryIds) {
 
 export async function fetchDpgStatuses(projectIds) {
   const { data: projectDpgStatuses, error } = await supabase
-    .from("project_dpg_status")
-    .select("project_id")
-    .in("project_id", projectIds);
+    .from('project_dpg_status')
+    .select('project_id')
+    .in('project_id', projectIds);
 
   if (error) throw new Error(error.message);
   return projectDpgStatuses;
 }
 
-export function mapProjectsWithTagsAndStatus(
-  projects,
-  projectCategories,
-  categories,
-  dpgStatuses
-) {
+export function mapProjectsWithTagsAndStatus(projects, projectCategories, categories, dpgStatuses) {
   const categoriesById = categories.reduce((acc, category) => {
     acc[category.id] = category;
     return acc;
@@ -54,7 +49,7 @@ export function mapProjectsWithTagsAndStatus(
     return {
       ...project,
       tags: tagsForProject.filter(Boolean),
-      dpgStatusCount: dpgStatusCountByProject[project.id] || "",
+      dpgStatusCount: dpgStatusCountByProject[project.id] || '',
     };
   });
 }
